backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql>=1.0.0
cryptography>=3.0.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
ortools==9.8.3296
pandas==2.1.3
openpyxl==3.1.2
jinja2==3.1.2
pydantic[email]==2.5.0
